Added a helper file for CustomerFormCard component.
